pyproject.toml
[project]
name = "google-gmail-tool"
version = "0.1.0"
description = "A CLI that provides access to google services like mail, calendar, drive"
requires-python = ">=3.14"
dependencies = [
    "click>=8.1.7",
    "google-api-python-client>=2.147.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "html2text>=2024.2.26",
    "python-slugify>=8.0.0",
    "python-dateutil>=2.8.2",
]
authors = [
    {name = "Dennis Vriend", email = "dvriend@ilionx.com"}
]
license = {text = "MIT"}
readme = "README.md"

[project.scripts]
google-gmail-tool = "google_gmail_tool.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.19",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
strict = true

[[tool.mypy.overrides]]
module = ["googleapiclient.*"]
ignore_missing_imports = true
